+++ lib/datamappify/data/criteria/active_record/order.rb
@@ -4,12 +4,12 @@ module Datamappify
   module Data
     module Criteria
       module ActiveRecord
-        class Where < CriteriaMethod
+        class Order < CriteriaMethod
           # @param scope [ActiveRecord::Relation]
           #
           # @return [ActiveRecord::Relation]
           def records(scope = nil)
-            records_scope(scope).where(structured_criteria)
+            records_scope(scope).order(structured_criteria)
           end
         end
       end